Refactored code into functions
